+++ packages/database/src/schemas/connector.ts
@@ -1,15 +1,17 @@
+import type { CustomPluginParams, ToolManifest } from '@lobechat/types';
 import {
   boolean,
   index,
   jsonb,
   pgTable,
+  primaryKey,
   text,
   uniqueIndex,
   uuid,
   varchar,
 } from 'drizzle-orm/pg-core';
 
-import { timestamps, timestamptz } from './_helpers';
+import { timestamps, timestamptz, varchar255 } from './_helpers';
 import { users } from './user';
 import { workspaces } from './workspace';
 
@@ -278,3 +280,27 @@ export const userConnectorTools = pgTable(
 
 export type NewUserConnectorTool = typeof userConnectorTools.$inferInsert;
 export type UserConnectorToolItem = typeof userConnectorTools.$inferSelect;
+
+export const userInstalledPlugins = pgTable(
+  'user_installed_plugins',
+  {
+    userId: text('user_id')
+      .references(() => users.id, { onDelete: 'cascade' })
+      .notNull(),
+    workspaceId: text('workspace_id').references(() => workspaces.id, { onDelete: 'cascade' }),
+
+    identifier: text('identifier').notNull(),
+    type: text('type', { enum: ['plugin', 'customPlugin'] }).notNull(),
+    manifest: jsonb('manifest').$type<ToolManifest>(),
+    settings: jsonb('settings'),
+    customParams: jsonb('custom_params').$type<CustomPluginParams>(),
+    source: varchar255('source'),
+    ...timestamps,
+  },
+  (self) => ({
+    id: primaryKey({ columns: [self.userId, self.identifier] }),
+  }),
+);
+
+export type NewInstalledPlugin = typeof userInstalledPlugins.$inferInsert;
+export type InstalledPluginItem = typeof userInstalledPlugins.$inferSelect;
